move struct, union, and message generation to their own files
